api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
python-multipart==0.0.6
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.0.1
scikit-learn==1.3.2
xgboost==2.0.1
passlib==1.7.4
python-jose==3.3.0
pyjwt==2.8.0
bcrypt==4.0.1
prompt-toolkit==3.0.39
fastapi-instrumentator>=0.4.0
psutil==5.9.6
gunicorn==21.2.0